misc. changes

+++ vntr.h
@@ -63,8 +63,8 @@ class VNTR
     float exact_entropy ;             //sequence entropy of repeat tract
     int32_t exact_beg1;               //beginning of repeat tract
     int32_t exact_end1;               //end of repeat tract
-    float exact_rl;                   //length of repeat tract in base pairs
-    float exact_ll;                   //length of repeat tract (including longest alternate allele) in base pairs
+    int32_t exact_rl;                   //length of repeat tract in base pairs
+    int32_t exact_ll;                   //length of repeat tract (including longest alternate allele) in base pairs
     int32_t exact_no_exact_ru;        //number exact repeat units from hmm
     int32_t exact_total_no_ru;        //total no of repeat units from hmm
     float exact_score;                //motif concordance from hmm
@@ -80,8 +80,8 @@ class VNTR
     float fuzzy_entropy ;             //sequence entropy of repeat tract
     int32_t fuzzy_beg1;               //beginning of repeat tract
     int32_t fuzzy_end1;               //end of repeat tract
-    float fuzzy_rl;                   //length of repeat tract in base pairs
-    float fuzzy_ll;                   //length of repeat tract (including longest alternate allele) in base pairs
+    int32_t fuzzy_rl;                   //length of repeat tract in base pairs
+    int32_t fuzzy_ll;                   //length of repeat tract (including longest alternate allele) in base pairs
     int32_t fuzzy_no_exact_ru;        //number exact repeat units from hmm
     int32_t fuzzy_total_no_ru;        //total no of repeat units from hmm
     float fuzzy_score;                //motif concordance from hmm
